% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isay.R
\name{isay}
\alias{isay}
\title{say hellp}
\usage{
isay(size = 1, ...)
}
\arguments{
\item{size}{number of time to greet}

\item{...}{other args passed to \code{\link[=sample]{sample()}}.}
}
\value{
a greeting
}
\description{
say hellp
}
